+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added automatic cleanup of media, when canceling outbox message
 - Added UriExtractor to `EventContentMediaMappings` (previously `OutboxMessageMediaUploaderMappings`)
+- download size limit for media service
+- default download limit properties in matrix client configuration
 
 ### Changed